Change weather/photo to always return images
Due to me not knowing what summaries the DarkSky API can return, there
is a issue with darksky sometimes giving a summary that isn't in the
allowdForecasts array, making the frontend app a lot less pretty. Until
someone with more knowlege on the API can help me with this I've removed
the filter (the API docs for DarkSky are now depreciated and no longer
exist...).